+++ app/build.gradle
@@ -41,8 +41,8 @@ dependencies {
     // EasyPermissions
     implementation 'pub.devrel:easypermissions:3.0.0'
     //Glide
-    implementation 'com.github.bumptech.glide:glide:4.9.0'
-    annotationProcessor 'com.github.bumptech.glide:compiler:4.9.0'
+    implementation 'com.github.bumptech.glide:glide:4.10.0'
+    annotationProcessor 'com.github.bumptech.glide:compiler:4.10.0'
     // ShapeOfView
     implementation 'com.github.florent37:shapeofview:1.3.2'
     // Support Design
@@ -62,6 +62,8 @@ dependencies {
     implementation 'com.squareup.retrofit2:converter-gson:2.6.0'
     // RxJava Call Adapter
     implementation "com.squareup.retrofit2:adapter-rxjava2:2.5.0"
+    // RxPermissions
+    implementation 'com.github.tbruyelle:rxpermissions:0.10.2'
     // ViewModel and LiveData
     //noinspection GradleDependency
     implementation "androidx.lifecycle:lifecycle-extensions:$lifecycle_version"
