--- app/src/main/java/com/assem/cognitev/nearby/Models/Photos/Photo.java
@@ -1,50 +1,0 @@
-
-package com.assem.cognitev.nearby.Models.Photos;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class Photo {
-
-    @SerializedName("prefix")
-    @Expose
-    private String prefix;
-    @SerializedName("suffix")
-    @Expose
-    private String suffix;
-
-    /**
-     * No args constructor for use in serialization
-     * 
-     */
-    public Photo() {
-    }
-
-    /**
-     * 
-     * @param prefix
-     * @param suffix
-     */
-    public Photo(String prefix, String suffix) {
-        super();
-        this.prefix = prefix;
-        this.suffix = suffix;
-    }
-
-    public String getPrefix() {
-        return prefix;
-    }
-
-    public void setPrefix(String prefix) {
-        this.prefix = prefix;
-    }
-
-    public String getSuffix() {
-        return suffix;
-    }
-
-    public void setSuffix(String suffix) {
-        this.suffix = suffix;
-    }
-
-}
