--- app/src/main/java/com/assem/cognitev/nearby/Models/Photos/Source.java
@@ -1,50 +1,0 @@
-
-package com.assem.cognitev.nearby.Models.Photos;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class Source {
-
-    @SerializedName("name")
-    @Expose
-    private String name;
-    @SerializedName("url")
-    @Expose
-    private String url;
-
-    /**
-     * No args constructor for use in serialization
-     * 
-     */
-    public Source() {
-    }
-
-    /**
-     * 
-     * @param name
-     * @param url
-     */
-    public Source(String name, String url) {
-        super();
-        this.name = name;
-        this.url = url;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getUrl() {
-        return url;
-    }
-
-    public void setUrl(String url) {
-        this.url = url;
-    }
-
-}
