--- app/src/main/java/com/assem/cognitev/nearby/Models/Photos/User.java
@@ -1,76 +1,0 @@
-
-package com.assem.cognitev.nearby.Models.Photos;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class User {
-
-    @SerializedName("id")
-    @Expose
-    private String id;
-    @SerializedName("firstName")
-    @Expose
-    private String firstName;
-    @SerializedName("lastName")
-    @Expose
-    private String lastName;
-    @SerializedName("photo")
-    @Expose
-    private Photo photo;
-
-    /**
-     * No args constructor for use in serialization
-     * 
-     */
-    public User() {
-    }
-
-    /**
-     * 
-     * @param firstName
-     * @param lastName
-     * @param photo
-     * @param id
-     */
-    public User(String id, String firstName, String lastName, Photo photo) {
-        super();
-        this.id = id;
-        this.firstName = firstName;
-        this.lastName = lastName;
-        this.photo = photo;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public Photo getPhoto() {
-        return photo;
-    }
-
-    public void setPhoto(Photo photo) {
-        this.photo = photo;
-    }
-
-}
