--- app/src/main/java/com/assem/cognitev/nearby/Models/Venues/Category.java
@@ -1,132 +1,0 @@
-
-package com.assem.cognitev.nearby.Models.Venues;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class Category {
-
-    @SerializedName("id")
-    @Expose
-    private String id;
-    @SerializedName("name")
-    @Expose
-    private String name;
-    @SerializedName("pluralName")
-    @Expose
-    private String pluralName;
-    @SerializedName("shortName")
-    @Expose
-    private String shortName;
-    @SerializedName("icon")
-    @Expose
-    private Icon icon;
-    @SerializedName("primary")
-    @Expose
-    private boolean primary;
-
-    /**
-     * No args constructor for use in serialization
-     * 
-     */
-    public Category() {
-    }
-
-    /**
-     * 
-     * @param pluralName
-     * @param name
-     * @param icon
-     * @param id
-     * @param shortName
-     * @param primary
-     */
-    public Category(String id, String name, String pluralName, String shortName, Icon icon, boolean primary) {
-        super();
-        this.id = id;
-        this.name = name;
-        this.pluralName = pluralName;
-        this.shortName = shortName;
-        this.icon = icon;
-        this.primary = primary;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public Category withId(String id) {
-        this.id = id;
-        return this;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public Category withName(String name) {
-        this.name = name;
-        return this;
-    }
-
-    public String getPluralName() {
-        return pluralName;
-    }
-
-    public void setPluralName(String pluralName) {
-        this.pluralName = pluralName;
-    }
-
-    public Category withPluralName(String pluralName) {
-        this.pluralName = pluralName;
-        return this;
-    }
-
-    public String getShortName() {
-        return shortName;
-    }
-
-    public void setShortName(String shortName) {
-        this.shortName = shortName;
-    }
-
-    public Category withShortName(String shortName) {
-        this.shortName = shortName;
-        return this;
-    }
-
-    public Icon getIcon() {
-        return icon;
-    }
-
-    public void setIcon(Icon icon) {
-        this.icon = icon;
-    }
-
-    public Category withIcon(Icon icon) {
-        this.icon = icon;
-        return this;
-    }
-
-    public boolean isPrimary() {
-        return primary;
-    }
-
-    public void setPrimary(boolean primary) {
-        this.primary = primary;
-    }
-
-    public Category withPrimary(boolean primary) {
-        this.primary = primary;
-        return this;
-    }
-
-}
