--- app/src/main/java/com/assem/cognitev/nearby/Models/Venues/Example.java
@@ -1,43 +1,0 @@
-
-package com.assem.cognitev.nearby.Models.Venues;
-
-import java.util.List;
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class Example {
-
-    @SerializedName("items")
-    @Expose
-    private List<Item> items = null;
-
-    /**
-     * No args constructor for use in serialization
-     * 
-     */
-    public Example() {
-    }
-
-    /**
-     * 
-     * @param items
-     */
-    public Example(List<Item> items) {
-        super();
-        this.items = items;
-    }
-
-    public List<Item> getItems() {
-        return items;
-    }
-
-    public void setItems(List<Item> items) {
-        this.items = items;
-    }
-
-    public Example withItems(List<Item> items) {
-        this.items = items;
-        return this;
-    }
-
-}
