--- app/src/main/java/com/assem/cognitev/nearby/Models/Venues/Icon.java
@@ -1,60 +1,0 @@
-
-package com.assem.cognitev.nearby.Models.Venues;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class Icon {
-
-    @SerializedName("prefix")
-    @Expose
-    private String prefix;
-    @SerializedName("suffix")
-    @Expose
-    private String suffix;
-
-    /**
-     * No args constructor for use in serialization
-     * 
-     */
-    public Icon() {
-    }
-
-    /**
-     * 
-     * @param prefix
-     * @param suffix
-     */
-    public Icon(String prefix, String suffix) {
-        super();
-        this.prefix = prefix;
-        this.suffix = suffix;
-    }
-
-    public String getPrefix() {
-        return prefix;
-    }
-
-    public void setPrefix(String prefix) {
-        this.prefix = prefix;
-    }
-
-    public Icon withPrefix(String prefix) {
-        this.prefix = prefix;
-        return this;
-    }
-
-    public String getSuffix() {
-        return suffix;
-    }
-
-    public void setSuffix(String suffix) {
-        this.suffix = suffix;
-    }
-
-    public Icon withSuffix(String suffix) {
-        this.suffix = suffix;
-        return this;
-    }
-
-}
