--- app/src/main/java/com/assem/cognitev/nearby/Models/Venues/Location.java
@@ -1,223 +1,0 @@
-
-package com.assem.cognitev.nearby.Models.Venues;
-
-import java.util.List;
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class Location {
-
-    @SerializedName("address")
-    @Expose
-    private String address;
-    @SerializedName("crossStreet")
-    @Expose
-    private String crossStreet;
-    @SerializedName("lat")
-    @Expose
-    private double lat;
-    @SerializedName("lng")
-    @Expose
-    private double lng;
-    @SerializedName("labeledLatLngs")
-    @Expose
-    private List<LabeledLatLng> labeledLatLngs = null;
-    @SerializedName("distance")
-    @Expose
-    private int distance;
-    @SerializedName("cc")
-    @Expose
-    private String cc;
-    @SerializedName("city")
-    @Expose
-    private String city;
-    @SerializedName("state")
-    @Expose
-    private String state;
-    @SerializedName("country")
-    @Expose
-    private String country;
-    @SerializedName("formattedAddress")
-    @Expose
-    private List<String> formattedAddress = null;
-
-    /**
-     * No args constructor for use in serialization
-     * 
-     */
-    public Location() {
-    }
-
-    /**
-     * 
-     * @param cc
-     * @param country
-     * @param address
-     * @param labeledLatLngs
-     * @param lng
-     * @param distance
-     * @param formattedAddress
-     * @param city
-     * @param state
-     * @param crossStreet
-     * @param lat
-     */
-    public Location(String address, String crossStreet, double lat, double lng, List<LabeledLatLng> labeledLatLngs, int distance, String cc, String city, String state, String country, List<String> formattedAddress) {
-        super();
-        this.address = address;
-        this.crossStreet = crossStreet;
-        this.lat = lat;
-        this.lng = lng;
-        this.labeledLatLngs = labeledLatLngs;
-        this.distance = distance;
-        this.cc = cc;
-        this.city = city;
-        this.state = state;
-        this.country = country;
-        this.formattedAddress = formattedAddress;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public Location withAddress(String address) {
-        this.address = address;
-        return this;
-    }
-
-    public String getCrossStreet() {
-        return crossStreet;
-    }
-
-    public void setCrossStreet(String crossStreet) {
-        this.crossStreet = crossStreet;
-    }
-
-    public Location withCrossStreet(String crossStreet) {
-        this.crossStreet = crossStreet;
-        return this;
-    }
-
-    public double getLat() {
-        return lat;
-    }
-
-    public void setLat(double lat) {
-        this.lat = lat;
-    }
-
-    public Location withLat(double lat) {
-        this.lat = lat;
-        return this;
-    }
-
-    public double getLng() {
-        return lng;
-    }
-
-    public void setLng(double lng) {
-        this.lng = lng;
-    }
-
-    public Location withLng(double lng) {
-        this.lng = lng;
-        return this;
-    }
-
-    public List<LabeledLatLng> getLabeledLatLngs() {
-        return labeledLatLngs;
-    }
-
-    public void setLabeledLatLngs(List<LabeledLatLng> labeledLatLngs) {
-        this.labeledLatLngs = labeledLatLngs;
-    }
-
-    public Location withLabeledLatLngs(List<LabeledLatLng> labeledLatLngs) {
-        this.labeledLatLngs = labeledLatLngs;
-        return this;
-    }
-
-    public int getDistance() {
-        return distance;
-    }
-
-    public void setDistance(int distance) {
-        this.distance = distance;
-    }
-
-    public Location withDistance(int distance) {
-        this.distance = distance;
-        return this;
-    }
-
-    public String getCc() {
-        return cc;
-    }
-
-    public void setCc(String cc) {
-        this.cc = cc;
-    }
-
-    public Location withCc(String cc) {
-        this.cc = cc;
-        return this;
-    }
-
-    public String getCity() {
-        return city;
-    }
-
-    public void setCity(String city) {
-        this.city = city;
-    }
-
-    public Location withCity(String city) {
-        this.city = city;
-        return this;
-    }
-
-    public String getState() {
-        return state;
-    }
-
-    public void setState(String state) {
-        this.state = state;
-    }
-
-    public Location withState(String state) {
-        this.state = state;
-        return this;
-    }
-
-    public String getCountry() {
-        return country;
-    }
-
-    public void setCountry(String country) {
-        this.country = country;
-    }
-
-    public Location withCountry(String country) {
-        this.country = country;
-        return this;
-    }
-
-    public List<String> getFormattedAddress() {
-        return formattedAddress;
-    }
-
-    public void setFormattedAddress(List<String> formattedAddress) {
-        this.formattedAddress = formattedAddress;
-    }
-
-    public Location withFormattedAddress(List<String> formattedAddress) {
-        this.formattedAddress = formattedAddress;
-        return this;
-    }
-
-}
