+++ app/src/main/java/com/assem/cognitev/nearby/Models/photos/Response.java
@@ -1,4 +1,4 @@
-package com.assem.cognitev.nearby.Models.Responses.photos;
+package com.assem.cognitev.nearby.Models.photos;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
