+++ app/src/main/java/com/assem/cognitev/nearby/Models/places/Place.java
@@ -1,9 +1,7 @@
-package com.assem.cognitev.nearby.Models.Responses.places;
-
-import android.widget.ImageView;
+package com.assem.cognitev.nearby.Models.places;
 
 
-import com.assem.cognitev.nearby.Models.Responses.photos.PhotoRespone;
+import com.assem.cognitev.nearby.Models.photos.PhotoRespone;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
