+++ app/src/main/java/com/assem/cognitev/nearby/UI/MainActivity.java
@@ -20,11 +20,13 @@ import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
 import androidx.core.widget.ContentLoadingProgressBar;
+import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProviders;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.assem.cognitev.nearby.App.MyApplication;
 import com.assem.cognitev.nearby.Helper.PrefManager;
+import com.assem.cognitev.nearby.Models.places.Item;
 import com.assem.cognitev.nearby.R;
 import com.assem.cognitev.nearby.Utils.BuildViews;
 import com.assem.cognitev.nearby.Utils.ConnectivityReceiver;
@@ -127,10 +129,13 @@ public class MainActivity extends AppCompatActivity
 
 
         // setup viewModel
-        venuesViewModel.itemsMutableLiveData.observe(this, items -> {
-            Log.d(TAG, "init: venue =>" + items.get(0).getVenue());
-            venuesAdapter.setList(items);
-            toggleLayout(true);
+        venuesViewModel.places.observe(this, new Observer<List<Item>>() {
+            @Override
+            public void onChanged(List<Item> items) {
+                Log.d(TAG, "init: venue =>" + items.get(0).getPlace().getName());
+                venuesAdapter.setList(items);
+                toggleLayout(true);
+            }
         });
         venuesViewModel.isEmptyMutableLiveData.observe(this, aBoolean -> {
             Log.d(TAG, "init: isEmpty => " + aBoolean);
@@ -177,9 +182,6 @@ public class MainActivity extends AppCompatActivity
 
     // call ViewModel actions
     private void getNearByVenues(Location location) {
-        venuesViewModel.getVenues(location);
-        venuesViewModel.getVenues_(location);
-        venuesViewModel.getVenuePhotos("55084193498eded0dc36c1e0");
         venuesViewModel.fetchPlaces(location);
 
     }
